VIDEO/THREAD: Fix 64-bit pointer truncation in texture handle pipeline
The threaded video wrapper's custom command pipeline used
32-bit types for texture handles:

  - custom_command_method_t returned int (32-bit)
  - custom_command.return_value was int (32-bit)
  - video_thread_texture_handle returned unsigned (32-bit)

On 64-bit platforms, texture handles are pointers (uintptr_t).
When a D3D9 COM pointer has upper 32 bits set (e.g.
0x0000020C4E7E2800), the return value is truncated to
0x4E7E2800.  The unload path then calls Release() on this
truncated pointer, crashing with an access violation.

This is the root cause of the D3D9 HLSL/Cg crash when running
with threaded video on 64-bit Windows.  It affects any driver
whose GPU allocations return addresses above 4GB.

Fix: change the entire pipeline to use uintptr_t:

  video_thread_wrapper.h:
    - custom_command_method_t returns uintptr_t
    - custom_command.return_value is uintptr_t
    - video_thread_texture_handle returns uintptr_t

  video_thread_wrapper.c:
    - Matching implementation change

  All video drivers + font_driver.c:
    - All _wrap callback functions return uintptr_t

14 files changed, 28 insertions, 28 deletions.

+++ gfx/drivers/d3d10.c
@@ -3666,7 +3666,7 @@ static void d3d10_gfx_unload_texture_internal(uintptr_t handle)
 }
 
 #ifdef HAVE_THREADS
-static int d3d10_texture_load_wrap(void *data)
+static uintptr_t d3d10_texture_load_wrap(void *data)
 {
    d3d10_texture_cmd_t *cmd = (d3d10_texture_cmd_t*)data;
    cmd->handle = d3d10_gfx_load_texture_internal(
@@ -3674,7 +3674,7 @@ static int d3d10_texture_load_wrap(void *data)
    return 0;
 }
 
-static int d3d10_texture_unload_wrap(void *data)
+static uintptr_t d3d10_texture_unload_wrap(void *data)
 {
    d3d10_texture_cmd_t *cmd = (d3d10_texture_cmd_t*)data;
    d3d10_gfx_unload_texture_internal(cmd->handle);
